+++ lifelines/fitters/coxph_fitter.py
@@ -71,8 +71,8 @@ class CoxPHFitter(RegressionFitter, ProportionalHazardMixin):
         See http://courses.washington.edu/b515/l17.pdf.
 
       n_baseline_knots: int
-        Used when ``baseline_estimation_method="spline"`. Set the number of _interior_ knots in the baseline hazard. Should be atleast 1. Royston et. al, the authors
-        of this model, suggest 2 to start, but any values between 1 and 4 are reasonable.
+        Used when ``baseline_estimation_method="spline"`. Set the number of knots (interior & exterior) in the baseline hazard. Should be atleast 2. Royston et. al, the authors
+        of this model, suggest 4 to start, but any values between 2 and 6 are reasonable.
 
     Examples
     --------
@@ -318,14 +318,8 @@ class CoxPHFitter(RegressionFitter, ProportionalHazardMixin):
 
         df = args[0].copy()
 
-        # handle if they provided a formula
+        # handle if they provided a formula or not
         formula = kwargs.pop("formula")
-        if type(formula) == str:  # TODO: change me
-            # expand like a formula with an intercept, and then drop the intercept.
-            # see https://patsy.readthedocs.io/en/latest/R-comparison.html
-            formula = "%s + 1" % formula
-        elif formula is None:
-            df["Intercept"] = 1
 
         # handle cluster_col
         if kwargs["cluster_col"] is not None:
@@ -341,7 +335,7 @@ class CoxPHFitter(RegressionFitter, ProportionalHazardMixin):
         strata = kwargs.pop("strata")
 
         if strata is None:
-            regressors = {**{"beta_": formula}, **{"phi%d_" % i: "1" for i in range(1, self.n_baseline_knots + 2)}}
+            regressors = {**{"beta_": formula}, **{"phi%d_" % i: "1" for i in range(1, self.n_baseline_knots + 1)}}
             strata_values = None
         elif isinstance(strata, (list, str)):
             spline_namer = ParametricSplinePHFitter._strata_spline_labeler
@@ -353,7 +347,7 @@ class CoxPHFitter(RegressionFitter, ProportionalHazardMixin):
             strata_values = df.groupby(strata).size().index.tolist()
             regressors = {"beta_": formula}
             for stratum in strata_values:
-                regressors.update({spline_namer(stratum, i): "1" for i in range(1, self.n_baseline_knots + 2)})
+                regressors.update({spline_namer(stratum, i): "1" for i in range(1, self.n_baseline_knots + 1)})
         else:
             raise ValueError("Wrong type for strata. String, None, or list of strings")
 
@@ -2396,14 +2390,15 @@ class ParametricSplinePHFitter(ParametricRegressionFitter, SplineFitterMixin, Pr
     _FAST_MEDIAN_PREDICT = False
 
     cluster_col = None
+    fit_intercept = True
 
     def __init__(self, strata, strata_values, n_baseline_knots=1, *args, **kwargs):
         self.strata = strata
         self.strata_values = strata_values
 
         assert (
-            n_baseline_knots is not None and n_baseline_knots > 0
-        ), "n_baseline_knots must be a positive integer. Set in class instantiation"
+            n_baseline_knots is not None and n_baseline_knots > 1
+        ), "n_baseline_knots should be greater than 1. Set in class instantiation"
 
         self.n_baseline_knots = n_baseline_knots
         super(ParametricSplinePHFitter, self).__init__(*args, **kwargs)
@@ -2417,13 +2412,13 @@ class ParametricSplinePHFitter(ParametricRegressionFitter, SplineFitterMixin, Pr
         if self.strata is not None:
             names = ["beta_"]
             for stratum in self.strata_values:
-                names += [self._strata_spline_labeler(stratum, i) for i in range(1, self.n_baseline_knots + 2)]
+                names += [self._strata_spline_labeler(stratum, i) for i in range(1, self.n_baseline_knots + 1)]
             return names
         else:
-            return ["beta_"] + ["phi%d_" % i for i in range(1, self.n_baseline_knots + 2)]
+            return ["beta_"] + ["phi%d_" % i for i in range(1, self.n_baseline_knots + 1)]
 
     def _set_knots(self, T, E):
-        self.knots = np.percentile(T[E.astype(bool).values], np.linspace(5, 95, self.n_baseline_knots + 2))
+        self.knots = np.percentile(T[E.astype(bool).values], np.linspace(5, 95, self.n_baseline_knots + 1))
         return
 
     def _pre_fit_model(self, Ts, E, df):
@@ -2441,7 +2436,7 @@ class ParametricSplinePHFitter(ParametricRegressionFitter, SplineFitterMixin, Pr
                     }
                 )
                 params.update(
-                    {self._strata_spline_labeler(stratum, i): np.array([0.0]) for i in range(3, self.n_baseline_knots + 2)}
+                    {self._strata_spline_labeler(stratum, i): np.array([0.0]) for i in range(3, self.n_baseline_knots + 1)}
                 )
 
             return params
@@ -2449,7 +2444,7 @@ class ParametricSplinePHFitter(ParametricRegressionFitter, SplineFitterMixin, Pr
         else:
             return {
                 **{"beta_": np.zeros(len(Xs["beta_"].columns)), "phi1_": np.array([0.05]), "phi2_": np.array([-0.05])},
-                **{"phi%d_" % i: np.array([0.0]) for i in range(3, self.n_baseline_knots + 2)},
+                **{"phi%d_" % i: np.array([0.0]) for i in range(3, self.n_baseline_knots + 1)},
             }
 
     def _cumulative_hazard_with_strata(self, params, T, Xs):
@@ -2470,7 +2465,7 @@ class ParametricSplinePHFitter(ParametricRegressionFitter, SplineFitterMixin, Pr
 
             H_ = safe_exp(anp.dot(Xs_["beta_"], params["beta_"]) + params[self._strata_spline_labeler(stratum, 1)] * lT_)
 
-            for i in range(2, self.n_baseline_knots + 2):
+            for i in range(2, self.n_baseline_knots + 1):
                 H_ = H_ * safe_exp(
                     params[self._strata_spline_labeler(stratum, i)]
                     * self.basis(lT_, anp.log(self.knots[i - 1]), anp.log(self.knots[0]), anp.log(self.knots[-1]))
@@ -2486,7 +2481,7 @@ class ParametricSplinePHFitter(ParametricRegressionFitter, SplineFitterMixin, Pr
 
         H = safe_exp(anp.dot(Xs["beta_"], params["beta_"]) + params["phi1_"] * lT)
 
-        for i in range(2, self.n_baseline_knots + 2):
+        for i in range(2, self.n_baseline_knots + 1):
             H = H * safe_exp(
                 params["phi%d_" % i] * self.basis(lT, anp.log(self.knots[i - 1]), anp.log(self.knots[0]), anp.log(self.knots[-1]))
             )
@@ -2578,7 +2573,6 @@ class ParametricSplinePHFitter(ParametricRegressionFitter, SplineFitterMixin, Pr
             df = df.to_frame().T.infer_objects()
 
         df = df.copy()
-        df["Intercept"] = 1
 
         if self.strata is not None:
             df = df.reset_index().set_index(self.strata)
@@ -2636,7 +2630,6 @@ class ParametricSplinePHFitter(ParametricRegressionFitter, SplineFitterMixin, Pr
             df = df.to_frame().T.infer_objects()
 
         df = df.copy()
-        df["Intercept"] = 1
 
         if self.strata is not None:
             df = df.reset_index().set_index(self.strata)
@@ -2663,11 +2656,6 @@ class ParametricSplinePHFitter(ParametricRegressionFitter, SplineFitterMixin, Pr
 
         else:
             return super(ParametricSplinePHFitter, self).predict_hazard(df, times=times, conditional_after=conditional_after)
-
-    def score(self, df: pd.DataFrame, scoring_method: str = "log_likelihood") -> float:
-        df = df.copy()
-        df["Intercept"] = 1
-        return super(ParametricSplinePHFitter, self).score(df, scoring_method)
 
     @property
     def AIC_partial_(self):