+++ perf_tests/cp_perf_test.py
@@ -16,8 +16,8 @@ if __name__ == "__main__":
     df = pd.concat([df] * reps)
     print(df.shape)
 
-    cph = CoxPHFitter(baseline_estimation_method="spline", n_baseline_knots=2, strata=["wexp"])
+    cph = CoxPHFitter(baseline_estimation_method="spline", n_baseline_knots=3, strata=["wexp"])
     start_time = time.time()
     cph.fit(df, duration_col="week", event_col="arrest", show_progress=True, timeline=np.linspace(1, 60, 100))
-    print(cph.baseline_hazard_)
+    print(cph.score(df))
     print("--- %s seconds ---" % (time.time() - start_time))